requirements.txt
aiogram==3.7.0
sqlalchemy==2.0.36
aiohttp==3.10.11
python-dotenv==1.0.1
aiosqlite==0.20.0
asyncpg==0.30.0
redis==5.2.1
pydantic==2.10.3